requirements.txt
tensorflow-directml
tensorflow-addons
notebook
pydot
ipykernel
matplotlib
pandas
imageio
graphviz
scipy
opencv-python
